+++ analysis/import/MetricGardenerImporter/src/test/kotlin/de/maibornwolff/codecharta/importer/metricgardenerimporter/MetricGardenerImporterTest.kt
@@ -21,8 +21,7 @@ class MetricGardenerImporterTest {
             return listOf(
                     Arguments.of("src/test/resources/my/supported-multi-language/repo"),
                     Arguments.of("src/test/resources/my/supported-multi-language/repo/dummyFile.js"),
-                    Arguments.of("src/test/resources/my"),
-                    Arguments.of(""))
+                    Arguments.of("src/test/resources/my"))
         }
 
         @JvmStatic
@@ -30,7 +29,8 @@ class MetricGardenerImporterTest {
             return listOf(
                     Arguments.of("src/test/resources/my/empty/repo"),
                     Arguments.of("src/test/resources/this/does/not/exist"),
-                    Arguments.of("src/test/resources/my/non-supported-language/repo"))
+                    Arguments.of("src/test/resources/my/non-supported-language/repo"),
+                    Arguments.of(""))
         }
     }