+++ analysis/parser/RawTextParser/src/test/kotlin/de/maibornwolff/codecharta/indentationlevelparser/RawTextParserTest.kt
@@ -25,8 +25,7 @@ class RawTextParserTest {
         fun provideValidInputFiles(): List<Arguments> {
             return listOf(
                     Arguments.of("src/test/resources/sampleproject"),
-                    Arguments.of("src/test/resources/sampleproject/tabs.xyz"),
-                    Arguments.of(""))
+                    Arguments.of("src/test/resources/sampleproject/tabs.xyz"))
         }
     }
 
@@ -105,6 +104,8 @@ class RawTextParserTest {
     fun `should NOT be identified as applicable if given directory does not contain any source code file `() {
         val emptyFolderPath = "src/test/resources/empty"
         val nonExistentPath = "src/test/resources/this/does/not/exist"
+        // Empty String is invalid because File("") generates an empty abstract pathname which does not physically exist
+        val emptyString = ""
 
         val emptyFolder = File(emptyFolderPath)
         emptyFolder.mkdir()
@@ -112,8 +113,10 @@ class RawTextParserTest {
 
         val isEmptyPathApplicable = RawTextParser().isApplicable(emptyFolderPath)
         val isNonExistentPathApplicable = RawTextParser().isApplicable(nonExistentPath)
+        val isEmptyStringApplicable = RawTextParser().isApplicable(emptyString)
 
         Assertions.assertThat(isEmptyPathApplicable).isFalse()
         Assertions.assertThat(isNonExistentPathApplicable).isFalse()
+        Assertions.assertThat(isEmptyStringApplicable).isFalse()
     }
 }